+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `context delete <PROFILE_NAME>` command to delete context profile.
+- Added `Network Impact` for `queries` command.
+
+### Deprecated
+- Remove ability sort by non impact columns in queries command.
 
 ## [0.2.1] - 2025-06-22